feature : 전략패턴 도입
