+++ src/main/java/com/company/finsight/api/article/service/ArticleService.java
@@ -5,7 +5,7 @@ import com.company.finsight.api.article.domain.Article;
 import com.company.finsight.api.article.dto.ArticleDetailDto;
 import com.company.finsight.api.article.dto.ArticleSummaryDto;
 import com.company.finsight.api.article.dto.ArticlesDto;
-import com.company.finsight.global.YNSCategory;
+import com.company.finsight.global.ArticleCategory;
 import com.company.finsight.global.exception.business.article.ArticleErrorCode;
 import com.company.finsight.global.exception.business.article.ArticleException;
 import com.company.finsight.api.article.repository.ArticleRepository;
@@ -30,7 +30,7 @@ import java.util.*;
 public class ArticleService {
 
     private final CrawlerClient crawlerClient;
-    private final ArticleParser articleParser;
+    private final ArticleParserFactory articleParserFactory;
     private final ArticleRepository articleRepository;
 
     /**
@@ -94,10 +94,10 @@ public class ArticleService {
         int MAX_CONCURRENCY = 5;
         long startTime = System.currentTimeMillis();
 
-        Flux.fromArray(YNSCategory.values())
+        Flux.fromArray(ArticleCategory.values())
                 .flatMap(category ->
                                 fetchCategoryArticleList(category)
-                                        .flatMap(this::fetchArticleContents)
+                                        .flatMap(articleList -> fetchArticleContents(articleList, category))
                                         .doOnError(error -> log.error("크롤링 실패 카테고리 : {}", category.getKoreanName(), error))
                                         .onErrorResume(e -> Mono.empty())
                         , MAX_CONCURRENCY)
@@ -111,23 +111,23 @@ public class ArticleService {
         log.info("스케줄링 종료.");
     }
 
-    private Mono<List<ArticleSummaryDto>> fetchCategoryArticleList(YNSCategory category) {
+    private Mono<List<ArticleSummaryDto>> fetchCategoryArticleList(ArticleCategory category) {
         return crawlerClient.call(category.getPath())
-                .flatMap(html -> articleParser.parseArticleList(html, category.getKoreanName(), crawlerClient.getYnsBaseUrl()))
+                .flatMap(html -> articleParserFactory.getParser(category.getArticleType()).parseArticleList(html, category.getKoreanName(), crawlerClient.getYnsBaseUrl()))
                 .doOnNext(articleList ->
                         log.info("파싱 결과 개수 : {}, 카테고리 : {}", articleList.size(), category.getKoreanName())
                 );
     }
 
-    private Mono<Void> fetchArticleContents(List<ArticleSummaryDto> articleList) {
+    private Mono<Void> fetchArticleContents(List<ArticleSummaryDto> articleList, ArticleCategory category) {
         Random random = new Random();
         int min = 5000;
         int max = 15000;
 
         return Flux.fromIterable(articleList)
-                .concatMap(articleSummary -> //
+                .concatMap(articleSummary ->
                         crawlerClient.call(articleSummary.getArticleUrl())
-                                .flatMap(articleParser::parseArticleContent)
+                                .flatMap(html -> articleParserFactory.getParser(category.getArticleType()).parseArticleContent(html))
                                 .publishOn(Schedulers.boundedElastic())
                                 .doOnNext(articleContent -> {
                                     log.info("본문 {} : 기자 {}", articleContent.getContent(), articleContent.getReporter());
