+++ src/main/java/com/company/finsight/api/article/service/ArticleType.java
@@ -1,0 +1,5 @@
+package com.company.finsight.api.article.service;
+
+public enum ArticleType {
+    YNS, HANKYUNG
+}
